# generated/csharp/src/Org.OpenAPITools/Controllers/TenantRolesApi.cs
/*
 * (title)
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TenantRolesApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="createRoleRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPost]
        [Route("/api/v1/tenants/{tenantId}/roles")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesCreateRole")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesCreateRole([FromRoute (Name = "tenantId")][Required]string tenantId, [FromBody]CreateRoleRequest createRoleRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"code\" : \"code\",\r\n  \"name\" : \"name\",\r\n  \"description\" : \"description\",\r\n  \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="roleId"></param>
        /// <response code="204">There is no content to send for this request, but the headers may be useful. </response>
        /// <response code="0">An unexpected error response.</response>
        [HttpDelete]
        [Route("/api/v1/tenants/{tenantId}/roles/{roleId}")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesDeleteRole")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesDeleteRole([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "roleId")][Required]string roleId)
        {

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="roleId"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpGet]
        [Route("/api/v1/tenants/{tenantId}/roles/{roleId}")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesGetRole")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesGetRole([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "roleId")][Required]string roleId)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"code\" : \"code\",\r\n  \"name\" : \"name\",\r\n  \"description\" : \"description\",\r\n  \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="page"></param>
        /// <param name="pageSize"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpGet]
        [Route("/api/v1/tenants/{tenantId}/roles")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesListRoles")]
        [SwaggerResponse(statusCode: 200, type: typeof(TenantRolesListRoles200Response), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesListRoles([FromRoute (Name = "tenantId")][Required]string tenantId, [FromQuery (Name = "page")]int? page, [FromQuery (Name = "pageSize")]int? pageSize)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"items\" : [ {\r\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"code\" : \"code\",\r\n    \"name\" : \"name\",\r\n    \"description\" : \"description\",\r\n    \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n    \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n  }, {\r\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"code\" : \"code\",\r\n    \"name\" : \"name\",\r\n    \"description\" : \"description\",\r\n    \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n    \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n  } ],\r\n  \"page\" : 0,\r\n  \"pageSize\" : 20,\r\n  \"total\" : 0\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<TenantRolesListRoles200Response>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="roleId"></param>
        /// <param name="tenantRolesSetPermissionsRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPut]
        [Route("/api/v1/tenants/{tenantId}/roles/{roleId}/permissions")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesSetPermissions")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesSetPermissions([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "roleId")][Required]string roleId, [FromBody]TenantRolesSetPermissionsRequest tenantRolesSetPermissionsRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"code\" : \"code\",\r\n  \"name\" : \"name\",\r\n  \"description\" : \"description\",\r\n  \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="roleId"></param>
        /// <param name="updateRoleRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPatch]
        [Route("/api/v1/tenants/{tenantId}/roles/{roleId}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantRolesUpdateRole")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantRolesUpdateRole([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "roleId")][Required]string roleId, [FromBody]UpdateRoleRequest updateRoleRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"code\" : \"code\",\r\n  \"name\" : \"name\",\r\n  \"description\" : \"description\",\r\n  \"permissionIds\" : [ \"permissionIds\", \"permissionIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}