# generated/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /api/v1/tenants/{tenantId}/roles:
    get:
      operationId: TenantRoles_listRoles
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - page
                  - pageSize
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
                  page:
                    type: integer
                    format: int32
                    example: 0
                  pageSize:
                    type: integer
                    format: int32
                    example: 20
                  total:
                    type: integer
                    format: int64
                    example: 0
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
    post:
      operationId: TenantRoles_createRole
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
  /api/v1/tenants/{tenantId}/roles/{roleId}:
    get:
      operationId: TenantRoles_getRole
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
    patch:
      operationId: TenantRoles_updateRole
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
    delete:
      operationId: TenantRoles_deleteRole
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
  /api/v1/tenants/{tenantId}/roles/{roleId}/permissions:
    put:
      operationId: TenantRoles_setPermissions
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissionIds:
                  type: array
                  items:
                    type: string
              required:
                - permissionIds
...
components:
  schemas:
...
    CreateRoleRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    Role:
      type: object
      required:
        - id
        - tenantId
        - code
        - name
        - permissionIds
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        code:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissionIds:
          type: array
          items:
            type: string
...
servers:
  - url: https://api.example.com
    description: Production
    variables: {}